[refactor]: UI component 파일 구조 변경
1. chekbox: index 추가
2. dataGrid: 폴더명 Pascal case 변경
3. datePicker: 폴더명 변경 및 type 분리
4. radioGroup: 폴더명 변경 및 type 분리
5. textField: 불필요한 파일 제거
