+++ src/shared/ui/checkbox/index.ts
@@ -1,0 +1,3 @@
+export { Checkbox} from '@/shared/ui/checkbox/Checkbox.tsx';
+
+export {options} from '@/shared/ui/checkbox/data.ts'
