+++ src/shared/ui/datePicker/DatePicker.tsx
@@ -1,16 +1,12 @@
 import { DatePicker as MuiDatePicker } from '@mui/x-date-pickers-pro';
-import type { DatePickerProps as MuiDatePickerProps } from '@mui/x-date-pickers-pro';
 import { AdapterDayjs } from '@mui/x-date-pickers-pro/AdapterDayjs';
 import { LocalizationProvider } from '@mui/x-date-pickers-pro/LocalizationProvider';
-import type { Dayjs } from 'dayjs';
 import dayjs from 'dayjs';
 import timezone from 'dayjs/plugin/timezone';
 import utc from 'dayjs/plugin/utc';
+import type { DatePickerProps } from './type';
 
-type DatePickerProps = MuiDatePickerProps<Dayjs> & {
-    size?: 'small' | 'medium';
-    fullWidth?: boolean;
-};
+
 dayjs.extend(utc);
 dayjs.extend(timezone);
 dayjs.tz.setDefault('Asia/Seoul');
