+++ src/shared/ui/datePicker/index.ts
@@ -1,0 +1,3 @@
+export * from '@/shared/ui/datePicker/DatePicker.tsx'
+
+export * from '@/shared/ui/datePicker/type.ts'
