+++ src/shared/ui/radioGroup/RadioGroup.tsx
@@ -1,7 +1,8 @@
 import { RadioGroupOptionData } from './data';
 import { FormControlLabel, RadioGroup as MuiRadioGroup, Radio } from '@mui/material';
+import type { RadioGroupProps } from './type';
 
-export const RadioGroup = ({ moduleName, ...props }) => {
+export const RadioGroup = ({ moduleName, ...props }: RadioGroupProps) => {
     const radioGroupOption = RadioGroupOptionData.find((option) => option.moduleName === moduleName);
     const radioGroupValue = radioGroupOption ? radioGroupOption.value : [];
