+++ src/shared/ui/radioGroup/data.ts
