+++ src/shared/ui/radioGroup/index.ts
@@ -1,0 +1,3 @@
+export * from '@/shared/ui/radioGroup/RadioGroup.tsx'
+
+export * from '@/shared/ui/radioGroup/data.ts'
