+++ src/shared/ui/radioGroup/type.ts
@@ -1,0 +1,5 @@
+import type { RadioGroupProps as MuiRadioGroupProps} from "@mui/material";
+
+export interface RadioGroupProps extends MuiRadioGroupProps {
+  moduleName: string;
+}
